+++ app/src/main/java/woowacourse/movie/model/ticket/seat/SeatToggleResult.kt
@@ -1,0 +1,11 @@
+package woowacourse.movie.model.ticket.seat
+
+sealed class SeatToggleResult {
+    data class Added(
+        val seat: Seat,
+    ) : SeatToggleResult()
+
+    data class Removed(
+        val seat: Seat,
+    ) : SeatToggleResult()
+}
